Корректирует файлы с модулями
